Fix: Address CodeRabbit review findings (round 3)
- Cap DPO trainer history to last 100 entries
- Resolve latest version in _create_deploy_symlink when version_id is None
- Cast samples_trained to int to avoid float
- Copy training_config before mutation in _build_training_section
- Document _validate_hyperparams behavior for missing values

+++ llm_learn/training/lora/trainer.py
@@ -237,7 +237,7 @@ class Trainer:
             },
             started_at=started_at,
             completed_at=utc_now(),
-            samples_trained=len(self.train_dataset) * self.training_config.num_epochs,  # type: ignore[arg-type]
+            samples_trained=int(len(self.train_dataset) * self.training_config.num_epochs),  # type: ignore[arg-type]
             adapter=adapter,
         )
